+++ pytorch_lightning/trainer/connectors/accelerator_connector.py
@@ -30,7 +30,7 @@ from pytorch_lightning.plugins import (
     DataParallelPlugin,
     DDP2Plugin,
     DDPFullyShardedStrategy,
-    DDPShardedPlugin,
+    DDPShardedStrategy,
     DDPSpawnPlugin,
     DDPSpawnShardedPlugin,
     DDPStrategy,
@@ -536,7 +536,7 @@ class AcceleratorConnector:
 
     @property
     def _is_sharded_training_type(self) -> bool:
-        return isinstance(self._training_type_plugin, (DDPShardedPlugin, DDPSpawnShardedPlugin))
+        return isinstance(self._training_type_plugin, (DDPShardedStrategy, DDPSpawnShardedPlugin))
 
     @property
     def _is_fully_sharded_training_type(self) -> bool:
@@ -722,7 +722,7 @@ class AcceleratorConnector:
             if use_tpu_spawn:
                 ddp_strategy_cls = TPUSpawnStrategy
             elif use_ddp_sharded:
-                ddp_strategy_cls = DDPShardedPlugin
+                ddp_strategy_cls = DDPShardedStrategy
             elif use_ddp_sharded_spawn:
                 ddp_strategy_cls = DDPSpawnShardedPlugin
             elif (